pyproject.toml
[project]
name = "test-task-django-effective-mobile"
version = "0.1.0"
description = ""
authors = [
    {name = "Dmitry",email = "kiselevdv1991@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2,<6.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "django-crispy-forms (>=2.4,<3.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "whitenoise (>=6.9.0,<7.0.0)",
    "drf-yasg (>=1.21.10,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-django = "^4.11.1"

